requirements.txt
discord
python-dotenv
pymongo
dnspython
lorem
pokepy
pyyaml
youtube-dl
flake8
black
